Update EnemyAttack_TopDown.cs
Don't love this way of managing enemies. It might be better to have a reload script which will take the original position of any enemies and then re-instantiate a new enemy there if the player dies? or find a way to just turn off enemies and disable their entire script if they die but then turn them back on and re-enable their script if they come back to life? That might be a better death