Adding layers of complexity
Trying to add too many layers of complexity all at the same time (want to have a room manager which will remember the state of the rooms even after you have left them (i.e. what doors are unlocked, if a chest has been opened) and I want to have doors that will pop you out at the right place and stuff like that, which is kinda complicated, especially because I am trying to do so many different things all at the same time, so I need to pick one thing and focus on it. I will probably make a room state controller which will determine if doors or gates are open or closed and will keep that consistent between loads