Update GameStatus_TopDown.cs
Added clarity to naming conventions. The game status works perfectly but only if it is not the original game status but the one stored in the dont destroy on load, so you need to either enter a room from a different room, or you need to reload that room and redo whatever it was you did before (i.e. die or kill all the enemies)