Using keyboard to control menus
Worked for the main and death menu but for some reason, it won't work with the pause menu. I think that the game is prioritizing controlling the player over the menu so inputs are turning the player and can cause them to buffer an attack, but can't control the menu. Will work on this eventually because I will eventually have the pause menu have the world map as well which I want the player to be able to look around