Main menu
Added an are you sure prompt for starting a new file and made the continue button disable if there is not currently a save file. All seems to work as intended (except for using keys to control the menu, but that is less of a priority right now).
Also added in keybinds for people who might prefer arrow keys over WASD, although I do not like them, might try to fix that all later on when I add controller compatibilty