Camera Controller seems to work moderately?
I want the camera to move with the player but only in specific situations (i.e. in a long vertical room, I want it to move up and down with it) soI made a controller and some triggers that should make that a bit more straight forward? Will hopefully also let me have some bigger and some smaller rooms?